first round of design adjustments
