+++ FRESHMusicPlayer/FRESHMusicPlayer-WPF/Pages/NotificationPage.xaml.cs
@@ -21,7 +21,7 @@ namespace FRESHMusicPlayer.Pages
         }
 
         private void InvalidateNotifications(object sender, EventArgs e) => ShowNotifications();
-        private void ShowNotifications()
+        private async void ShowNotifications()
         {
             NotificationList.Items.Clear();
             foreach (Notification box in window.NotificationHandler.Notifications)
@@ -29,7 +29,7 @@ namespace FRESHMusicPlayer.Pages
                 box.Read = true;
                 NotificationList.Items.Add(new NotificationBox(box, window.NotificationHandler));
             }
-            if (NotificationList.Items.Count == 0) (Application.Current.MainWindow as MainWindow)?.HideAuxilliaryPane();
+            if (NotificationList.Items.Count == 0) await window.HideAuxilliaryPane();
         }
 
         private void Page_Unloaded(object sender, RoutedEventArgs e)
